snowflake-embedding/requirements.txt
Flask
sentence-transformers==3.0.0